requirements.txt
streamlit
passlib>=1.7.4
bcrypt==3.2.2
pandas
matplotlib
psycopg2-binary